+++ src/hooks/ready/anchorInjection.js
@@ -105,11 +105,14 @@ export function anchorInjection() {
 
     // when we create from the side bar we fill in the input label name to match
     // the anchor name and set the slug value to the anchor slug
-    Hooks.once("renderNoteConfig", (noteConfig) => {
+    Hooks.once("renderNoteConfig", (noteConfig, _html, app) => {
       const titleInput = noteConfig.element[0].querySelector("input[name='text']");
       titleInput.setAttribute('value', dropData.anchor.name);
       if (dropData.anchor.slug) {
+        titleInput.setAttribute('value', dropData.anchor.name);
         updateNotePage(noteConfig, dropData.anchor.slug);
+      } else {
+        titleInput.setAttribute('value', app.label);
       }
     });
   });